Catch date format errors in control parser
This update makes some of the date parsing utilities top-level utils.
It causes the control parser to disregard all optional information
(date, authors, comments) if the date doesn’t follow the proper ISO8601
format. It sends a debug message instead of raising an exception like
other problems, because this is optional data. Closes #147