I'm hungry and this push is going to fail but its lunch time
@dexterpratt this commit solves the issue with the context entry not
looking like JSONLD. It also updates the names for each node to be
namespace:name pairs when available, or BEL statements when they’re
not. I checked uploading to NDEx and it yelled at me because it didn’t
have a status, so I added that on line 275 as well. The from_cx
functions are not yet updated to retrieve from this new format, but
maybe you can give some feedback before I put time into that